Changed license from MIT to GPL2
Changed license from MIT to GPL2 in order to be compatible with
Zenphoto’s GPL2 license.

I’m not totally clear from reading online whether the GPL2 license is
actually required, but there does seem to be precedent that the PHP
code in Wordpress plugins is considered to be “one program” with
Wordpress and thus must be released under GPL2 because Wordpress is.

For the record I would prefer this code to be freely available to
anyone to do whatever they want with it, including using it in
proprietary software — which is something GPL2 prohibits.  But I don’t
care enough to research it thoroughly.

This closes #3