Test perf of getting stats
Added a test that gets the gallery and a stat without getting any
albums (i.e. depth=0).  This was to compare the performance of this
against getting stats *with* albums, for the scenario where you just
care about getting some stats.

It isn’t a very meaningful comparison because the order tests are run
in affects database caching and other things.  I ran them 10 times and
took the average of the 10 runs:
128.8ms - Gallery depth 0
124.1ms - Gallery depth 0 with 1 stat
140.1ms - Gallery depth 1
148.9ms - Gallery depth 1 with 1 stat

You can see it’s not very meaningful: getting 1 stat with the gallery
is faster than just getting the gallery!

However, I can say that Gallery depth 0 seems to be the cheapest call
the API can make.  It’s a few ms faster than retrieving an image, 10 or
more ms faster than retrieving an album, and 20ms and up faster than
retrieving more than one level of albums.  So as of now I’m going to
recommend gallery depth 0 as the base for statistics-only calls.